Moved parkRobot, goThroughTrussAndFinish, and planPurple to respective classes. Made some deletions first committed in the cleanup branch.
